+++ src/components/Navbar.tsx
@@ -1,52 +1,10 @@
 'use client';
 
-import { usePathname } from 'next/navigation';
-import * as NavigationMenu from '@radix-ui/react-navigation-menu';
-import NextLink from 'next/link';
-import React, { ReactNode } from 'react';
 import Hamburger from './icons/Hamburger';
 import { SignedIn, UserButton, SignOutButton } from '@clerk/nextjs';
-import classNames from 'classnames';
-import { CaretDownIcon } from '@radix-ui/react-icons';
-import './navlarge.css';
+import NavLarge from './NavLarge';
 
-const Link = ({ href, ...props }: {children: string, href: string}) => {
-  const pathname = usePathname();
-  const isActive = href === pathname;
 
-  return (
-    <NavigationMenu.Link asChild active={isActive}>
-      <NextLink href={href} className="NavigationMenuLink" {...props} />
-    </NavigationMenu.Link>
-  );
-};
- 
-
-// ListItems in Nav Dropdown
-type ListItemType = { 
-  children?: ReactNode;
-  className?: string; 
-  href?: string; 
-  rel?: string; 
-  target?: string; 
-  title?: string; 
-}
-
-const ListItem = 
-  ({ className, rel, target, title, ...props }: ListItemType) => (
-    <li>
-      <NavigationMenu.Link asChild>
-        <a
-          className={classNames('ListItemLink', className)}
-          rel={rel}
-          target={target}
-          {...props}
-        >
-          <div className="ListItemHeading">{title}</div>
-        </a>
-      </NavigationMenu.Link>
-    </li>
-);
 
 // const Navbar = () => {
 //   return (
@@ -174,71 +132,11 @@ const ListItem =
 
 const Navbar = () => {
   return (
-    <header>
-      <div className='hidden md:flex'>
-        <NavigationMenu.Root className="NavigationMenuRoot">
-          <NavigationMenu.List className="NavigationMenuList">
-
-            <NavigationMenu.Item>
-              <Link href="/">Home</Link>
-            </NavigationMenu.Item>
-
-            <NavigationMenu.Item>
-              <Link href="/#history">About Juneteenth</Link>
-            </NavigationMenu.Item>
-
-            <NavigationMenu.Item>
-              <Link href="/schedule">Schedule</Link>
-            </NavigationMenu.Item>
-          
-            <NavigationMenu.Item>
-              <NavigationMenu.Trigger className="NavigationMenuTrigger">
-                Maps <CaretDownIcon className="CaretDown" aria-hidden />
-              </NavigationMenu.Trigger>
-              <NavigationMenu.Content className="NavigationMenuContent">
-                <ul className="List">
-                  <ListItem
-                    title="Juneteenth"
-                    href="downloads/juneteenth-map.pdf"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                  />
-                  <ListItem
-                    title="Youthfest"
-                    href="/downloads/youth-fest-map.pdf"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                  />
-                </ul>
-              </NavigationMenu.Content>
-            </NavigationMenu.Item>
-
-            <NavigationMenu.Item>
-              <Link href="/vendor-application">Vendor Signup</Link>
-            </NavigationMenu.Item>
-
-            <NavigationMenu.Item>
-              <Link href="/sponsors">Sponsors</Link>
-            </NavigationMenu.Item>
-        
-            <SignedIn>
-              <NavigationMenu.Item>
-                <NavigationMenu.Link>
-                  <SignOutButton />
-                </NavigationMenu.Link>
-              </NavigationMenu.Item>
-            </SignedIn>
-
-            <NavigationMenu.Indicator className="NavigationMenuIndicator">
-              <div className="Arrow" />
-            </NavigationMenu.Indicator>
-          </NavigationMenu.List>
-
-          <div className="ViewportPosition">
-            <NavigationMenu.Viewport className="NavigationMenuViewport" />
-          </div>
-        </NavigationMenu.Root>
+    <header className='bg-white py-1'>
+      <div className='md:hidden flex'>
+        <h3>small nav here</h3>
       </div>
+     <NavLarge/>
     </header>
   );
 };
